refactor: apply review. make dto in controller layer. repackaging
