+++ src/main/java/ladder/model/LadderHeight.java
@@ -1,4 +1,4 @@
-package ladder;
+package ladder.model;
 
 import java.util.Objects;
