+++ src/main/java/ladder/model/LadderLine.java
@@ -1,4 +1,4 @@
-package ladder;
+package ladder.model;
 
 import java.util.*;
 import java.util.stream.Collectors;
