+++ src/main/java/ladder/model/Members.java
@@ -1,4 +1,4 @@
-package ladder;
+package ladder.model;
 
 import java.util.Arrays;
 import java.util.Collections;
