+++ src/test/java/ladder/LadderBridgeTests.java
@@ -1,5 +1,6 @@
 package ladder;
 
+import ladder.model.LadderBridge;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
