+++ src/test/java/ladder/LadderHeightTests.java
@@ -1,5 +1,6 @@
 package ladder;
 
+import ladder.model.LadderHeight;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
